+++ pkg/conn/mysql_client.go
@@ -1,4 +1,4 @@
-package repo
+package conn
 
 import (
 	"database/sql"
